requirements.txt
Django
djangorestframework
pykafka
facebook-scraper